air, be/C: Fix compilation of pre-conditions for intrinsics, fix #264
Intrinsic features require their inner fields since they may be accessed (or
even defined) within the feature's pre-condition, which is compiled by the
backend and not treated as an intrinsic.

For the same reason, the C backend has to declare types for the instances of
intrinsics since these types may be used in the code generated for the
pre-condition.