fuir/analysis, be/C: Add AbstractInterpreter, extract code from C backend
A new class AbstractInterpreter provides the basic infrastructure for abstract
interpretation. In particular, the handling of the stack that should be common
to all abstract interpreter, whether for analysis or for code generation, has
been removed completely from the C backend.

Next, the AbstractInterpreter class should be used by DFA.