fuir/analysis: Improved DFA analysis
DFA analysis now iterates the Calls, not the Instances. Added support for const
strings and limited support for 'fuzion.sys.array'.

Added workaround to for outer ref to an outer unit type, which is read but never
written.

Added detection of diverging and non-diverging routines. At the beginning of the
analysis, any routine is diverging. Only if analysis finds a path that produces
a non-void result, the routine is marked to be non-divering and may produce a
result.

Added tracking of the context why a call or instance ends up in the DFA, a call
to 'showWhy' shows the reason recursively.