DFA: Fix thrown Error when match matches different types at once
This occurred in tests/typeinference.

The old code tried to join the different values, which fails if they have
different types.  The joined value was only needed when the match case assigned
it to a field, but then there could be only one value and joining would not
occur.

Simplified to the code to set the field's value within the lambda processing the
case tags so the untagged values are no longer needed after the lambda.