fuir, be/C: Fix 'match' if case actual generic matches more than type
This problem occurs when running idiom225ex.fz using the C backend: Here, a
match is done on 'option T' where T is 'list string' where 'nil' is assignable
to both the 'nil'-option and the to 'T' where 'T' is 'list string' and 'nil' is
the empty list.

The fix is to make the match cases depend only on the type assignability in the
front end, not on the actual clazzes used in FUIR.

This fixes the original and small examples from #304, but the new example still
does not work.