fe: Fixed visibility of re-assigned fields
In the following code that declares two fields with name 'x'

  x := 42
  f is
    say x
  q := x
  x := "hello"
  g is
    say x
  r := x

The lookup of 'x' occurs in the code of 'f' and 'g' and in the initialzer of
'q' and 'r'. The initializer case was not covered resulting in errors.

With this patch, when looking up fields, initial values are taken into
account.