fz tool: Enable saving of module files other than 'base.fum'
The actual saving does not work yet, but this provides the user interface
necessary to do this:

Option '-XsaveBaseLib' has been changed to '-saveLib'.

Added option '-XloadBaseLib=(on|off)' has been added to disable loading
'base.fum', which is required when 'base.fum' is created.

Added optin '-sourceDirs={'

Made 'saveLib' a proper backend, moved code to save the library from
FrontEnd.java to this backend.  Added code to handle an argument to the backend
option, in this case '-saveLib=<path>'.