fe/air/be: Implement type.name for C backend, fix #457
Cleanup the way instances of 'Type' are created and handled: Instance no longer
have a field 'name', but an intrinsic feature 'name'. So types are unit types
now without any data stored in inner fields.

Support for 'Type.name' intrinsic was added to all backends. The result is a
cleaner handling of type names, moving code for this to one remaining ugly part
within package 'air' during Clazzes.findClazzes where a type clazz's type name
is set from the type parameter that returns it when called (see accesses to
field Clazz._typeName).  It would be better to have this information available
when the Clazz instance corresponding to a type is created.