parser: Disallow whitespace in 'x y.type' and 'x y.env', fix #455
'x y.type' is now parsed as a call to 'x' with argument 'y.type', while 'x
y.env' is parsed as a call to 'x' with argument 'y.env'.

To get the old behaviour, you need to write '(x y).type' to the the type
instance of 'x y' (Type 'x' with type parameter 'y' is in 'list i32') or '(x
y).env'.