air: Add support for 'select' when looking up field result clazz, fix #456
The missing support for 'select' in the call to 'lookup' was the reason
'actualClazz0' had to be used instead, but the use of 'actualClazz0' instead of
a proper 'lookup' resulted in wrong behaviour seen in #456.

Now, with the call to 'lookup' fixed for 'select >= 0' the hack using
'actualClazz0' can be removed.