air: cleanup: Simplify code of 'Clazz.clazzForFieldX'
The special handling for 'field.isTypeParameter()' is not needed, and since
'field.outer()' is used only for fields that are outer refs there is also no
need to check that this is non-null, would be very strange if it was not.

With this change, the only remaining special case in this code compared to
'Clazz.lookup' is the type of an outer ref field that is the address of an outer
value type.