air: change clazz created by Clazz.typeClazz() to be a Clazz of value type
A type clazz is not only a value type, but a unit type, so passing a type around
is a NOP, intrinsics to create a type value do not need to do anything if unit
types are optimized out.

This cleanup also fixes module file support for nested types such as a module
file referencing a type 'fuzion.java.type' would result in two types, one whose
outer clazz is 'fuzion.#type' and one whose outer clazz is 'ref fuzion.#type',
resulting in bugs since the latter is never found to be instantiated.