fe: Enable loading of inner features from directory if outer is in .fum file
This enables, e.g., a source file 'string/countSpaces.fz' to add a feature
'counterSpaces' to 'string' which is loaded from 'base.fum'.

Note that currently, source files in sub-directories are loaded lazily after the
corresponding outer feature was loaded. This means that features in
sub-directories will not be added if the corresponding outer features are not
touched, e.g., if we built a module file, these might not get added.