tools/fz: Attempt for cleanup in usage output.
The idea is not to output the complete usage on -help, but to show the list of
backend options and encourage the user to type, e..g, 'fz -c -help' to get the
usage specific to a given backend ('-c' in this case).

As next steps, there should be backends added for '-version', '-pretty', etc. to
structure this further using backends.