Makefile: build most Java modules
This does not include the following modules that are included in Java
17:

- jdk.hotspot.agent
- jdk.jcmd
- jdk.incubator.concurrent
- jdk.incubator.vector

The former two only provide Java tools, the latter two provide
experimental features that can move in future Java releases.
Additionally, these modules cause obscure build errors.

Secondly, the following modules are not build automatically right now:

- java.xml.crypto
- java.se
- jdk.xml.dom

The two XML-related modules for some reason do still generate duplicate
features that are *not* removed by the changes that is supposed to fix
that. This needs to be investigated. The java.se module depends on
java.xml.crypto, for this reason it is disabled as well.