lib: Remove 'lib/java*' and add explicit dependency to JavaObject.type
JavaObject.type is used by Java related intrinsics, so the middle end does not
see that this is needed.  Adding this explicitly makes sure this is found.

Another (better) solution might be to add a mechanism to the middle end to add
dependencies that depend on intrinsics that are called.