tools/fzjava: cross-check modules in recurseDeclaredFeatures
It is not sufficient to start with the universe in one module and then
descend into it without considering features declared by other modules,
if the intent, as it is here, is to discover all the features declared
by the given module. Fix this, which in turn fixes the remaining
duplicate features.