fe/lib: Add Type_STATIC as direct parent of type features
The type feature f.type of a feature f now inherits from Type_STATIC and f's
non-static type feature, which itself inherits from all of f's parents' type
features:

  f : g, h

then

                Type
                  A
                  |
            Object.type (dyn)
                A  A
                |  |
     +----------+  +---------+            Type_STATIC
     |                       |               A
  g.type (dyn)            h.type(dyn)        |
     A                       A               |
     |                       |               |
     +--------+   +----------+               |
              |   |                          |
            f.type (dyn)                     |
                A                            |
                |                            |
                +-------+  +-----------------+
                        |  |
                      f.type (static)

Not shown are g.type (static) and h.type (static) that inherit from g.type (dyn)
/ h.type (dyn), resp., and Type_STATIC similar to f.type (static).

Type_STATIC now defines an abstract equality operation for two instances of that
type. string is currently the only feature implementing this.

A new feature 'equals' and a shortcut 'infix ≟' with a type parameter provide a
convenient way to call equality on that type.

A test for using equality has been added.

The workaround for #461 in AstErrors was actually not related to that issue and
has been removed.

There are still a number of issues with this patch that are marked with NYI:

 - a check in the interpreter fails since types' outer refs are used somewhere
   (but never initialized).

 - the 'infix ≟' cannot really be called since its left hand side is universe

 - parts of tests/equals are commented out since they run into an endless loop
   or (if enabled) a check failure.