fe: for operator calles, look for called feature in outer features
For an infix, prefix or postfix operator call of the form

  a ⨁ b     -- or --
  ⨁ a       -- or --
  a ⨁

that was not found within the target 'a', try to find this operator in 'this'
or any outer feature.  If found in X.Y.Z.this, then convert this call into

  X.Y.Z.this.infix  ⨁ a b     -- or --
  X.Y.Z.this.prefix ⨁ a       -- or --
  X.Y.Z.this.postix ⨁ a       ,

respectively.  This permits the introduction of binary or unary operators
within any feature, e.g., within unit type features that can be inherited
from or even in the universe.