tools/fzjava: when building multiple modules, separate them
Previously, fzjava would put the Fuzion features for all given modules
in a single directory. Now, if more than one module is given, we put the
Fuzion features for a Java module x into dest/x, for a Java module y
into dest/y, and so on, where dest is the destination directory given by
the -to command line argument.

We do not address the case where the given modules may depend on each
other, and hence might contain duplicate features afterwards. In this
case, building the first module as a single module, then compiling it
into a fum file, and then loading the fum file with -modules is still
necessary to avoid duplicate features.

Closes #599.