lib/fe: Add 'dyn' keyword for non-static type features
A type feature declared as 'dyn' will be added to the non-static type feature,
i.e., all type features of heir features will inherit it even if it is not
abstract.

This permits implementing something like 'numerics.from_u32' in
'numeric.type.from_u32'.  However, this does not work yet since 'numeric.T' is
incompatible with 'numeric.type.THIS_TYPE', so operations are not possible yet.

This is one small step towards solving #706.