lib/ast/fe: Support for implicit `a.b.c.this.type` (improving #167 and #706)
There is no support for explicit use of `a.b.c.this.type` yet, but this patch
defines a new flavour of types to be used for `a.b.c.this` and types inferred
from this expression.

A new type variant `isThisType` in addition to `isRef` and `isValue` has been
added for types that can be any heir of the underlying feature's type.

`Type.isAssignableFrom(Expr)` has been removed since the assingability for
`a.b.c.this` can now be checked using the new isThisType variant
`a.b.c.this.type`.

In the base library, fixes where needed in `outcome.and` (since `outcome.this`
is not assignment compatible to the result) and `list.tails` since the result
type for `head` must be `list A` and not `list.this.type`, which could be an
heir (but also it could not since you cannot inherit from a choice feature...).

There are some NYI: CLEANUP: comments in this patch for which I will add new
issues.