docs: detect code in comments and convert it to runcode box
A block of code is being detected in a comment when there is a continous
block of lines that are indented five spaces from the # (or four spaces
from where the text begins, when you separate the # and the text with a
single space). The code is then inserted into a runcode wrapper which
will be displayed on flang.dev as a box containing executable and
modifiable code.

Currently, this is only useful when the docs generated by fzdocs are
being included on flang.dev, as the necessary CSS and JavaScript stuff
is missing otherwise.

Link: https://git.tokiwa.software/tokiwa/flang_dev/issues/110