fe: Support 'a.this.type' as type for args/results with covariant redefinition, add 'fixed' features
The parser method for rule 'qualThis' has been modified to procude a
'List<String>' or 'Type' depending on where it is used.

Type 'a.this.type' is can be used in an inner feature of 'a' only for
assginments from any value that is also 'a.this.type' or that is 'a.this'. In an
heir feature 'b' that inherits from 'a', redefinitions must replace
'a.this.type' by 'b.this.type'.

Modifier 'fixed' can be used to mark a feature that will not be inherited by
heirs.  This is useful in 'b' that inherits from 'a' to provide an
implementation of an inherited feature with argument or result type
'a.this.type' since this type can then be replaced by 'b'.

Fixed equal value for dev.flang.ast.Const.MODIFIER_FIXED and MODIFIER_DYN.
Added MODIFIER_FIXED to module file.

test/thisType now checks more cases: q.this for q being ref and value type,
redef for fixed method,