lib: io: implement get_file_size as a fuzion feature
Previously, get_file_size was implemented as an intrinsic which, in the
C backend, called to stat. So does the stats intrinsic, therefore we can
actually move to using that.

Also mark get_file_size private for now. It uses the stats intrinsic,
but without using the stat effect. We might want to change this. Keep it
private until a decision is made.