fe/lib/tests: Use 'fixed' for concrete type feature implementations
Instead of distinguishind type features that are 'dyn'amic and 'static' using
two different classes, this patch now ueses 'fixed' to mark type feature
implementations that cannot be inherited by type features of children.

This will, in a separate PR, enable the removal of the 'dyn' keyword and the removal of static type features.