fuir: Check that Clazzes.error.get() is not part for the clazzes in FUIR
This may only occur as a consequence of earlier errors and a bug that would
result in a pre- or postcondition failure.

However, when pre/postconditios are disabled, generating C code for the error
clazz is always confusing, so we better do not try to work with it at all.