lib/fe/air/tests: 'a.this.type' to get actual type of outer feature a
This enables accessing the actual type of an outer 'this' instance and allows
calling features redefined for a specific type.

This is expected to make the generic parameter in features like 'numeric'
obsolete, this parameter can be replaced by 'numeric.this.type'.

Extended tests/covariance with test cases that check different levels of outer
this.types being accessed and redefined.