ast: Support for type parameters in type features
Type features now have copies of the type parameters of their underlying
original features in addition to THIS#TYPE.  As an example, the type feature
option.type, where

  option(T type)

has one type parameters, has two:

  option.type(THIS#TYPE type, T type)

An the type of, e.g.

  option (list i32)

is

  option.type (option (list i32)) (list i32)

i.e., 'THIS#TYPE' is 'option (list i32)', while 'T' is 'list i32'

Special handling was needed for the case when actual type parameters are
inferred from actual arguments in a call. In this case, the type feature's does
not have the actual types ready when it is created, so a call-back is installed
to add this information when it becomes available.

Also, a special handling is needed for types that are cloned for use in type
features.  The clone will use its original context to look up features, but will
take the type feature to look up type parameters.

Also added a test case for type features with type parameters.