air: Mark clazzes of constants as instantiated, fix #886, enhance #874
The example of #886 created constants of type `u8` without performing any
calculations that produces an `u8`. Since the constants did falsely not mark
clazz `u8` as instantiated, the middle end assumed there were no instances and
hence no features could be called on `u8` resulting in this error when DFA
analysed a call to `u8.infix <`.

Now, the presence of a constant ensures that the clazz of the constant is marked
as instantiated.