lib/ast/tests: cleanup and add tests of type inference using this, fix #874
Removed post-processing of type for type inference, this turned out not to work
in all cases in the new test added in this patch.  Expr.inferredType() has been
removed, Expr.typeIfKnown can be used directly instead.

Whenever type inference is used with `this`, the type is now the corresponding
`this.type`. This required changes in `String.fz` that used `String.this` for
type inference resulting in a type incompatible with `String`.

Added tests/reg_issue874ff that contains a collection of test cases related to
type inference and this.type.