air: add expermimental code to eargerly replace 'a.this.type'
This is experimental: Clazz.actualGenerics() could replace `a.this.type` by the
actual outer type. This would avoid special handling in Clazzes.clazz(Expr e,
Clazz outerClazz). However, this is currently disabled since it has the
side-effect of marking some clazzes as instantiated that are actually not (in
tests/reg_issues874ff), need to check why.