lib/tests: add new feature local_mutate
This is work in progress with many NYI-comments, but local mutatability is
taking shape.

A new feature `local_mutate` provides locally mutable variables and locally
mutable arrays.

Eventually, it would be best to merge `local_mutate` into `mutate` and have only
one effect for mutability.  But I would like to fix all the errors we still have
with the current code first.

Added an example that uses local mutability for a local mutable `sum`, that uses
quicksort on a locally mutable array and that creates cyclic data structures
using local mutability in a way that these cycles can be frozen and then used in
pure code without requiring a mutate effect.