util/ast: Made List freezable, freeze generics list when used as key, fix #979
The bug that causes #979 was that AbstractType.actualType() used caching, while
the key used for the cache was a list of generic parameters was modified
afterwards, such that the cache contained a wrong result.

This patch adds methods to List to protect the List from modifications and also
ways to clone a list on demand when seting one or mapping all elements.