fe/air: Don't lookup internal features that were added late, fix #978
Internal features added for, e.g., lambda expressions do not show up in the set
od declared or inherited features of the children of their outer features.  This
is why a lookup would fail and we avoid this here.

This fix might need to be revisted when working on #932, for now this is just a
workaround.