lib/tests: Add `local_mutate.mget` and suggested syntax for effect in signature
`mget` has to be used while a mutable value is still mutable, while `get` can
only be used when it is closed.  `get` can easily be analysed statically not to
require a local_mutate effect to be present.

Also added possible syntax to create alias for types.

Updated tests/local_mutate to use mget.