air/lib/tests: Support this.type.env to access effect in environment
Using `a.this.type.env` automatically results in accessing the effect of a child
of the effect `a`.

This change actually adds support for `this.type` to all uses of
Clazz.actualClazz, but this seems to be only relevant to `this.type.env` for
now.

Updated base lib's `local_mutate` to no longer use a type parameter, but use
`local_mutate.this.type` instead.

Changed tests/local_mutate to no longer use these type parameters.