ast: Improve error created when arguments have equal names
This error was helpful when working on type features where I ran into a problem
due to repeated arguments.  But I needed more information to understand the
problem, so this patch.

Moved the check for this error to the end of the Feature constructor such that
all fields in feature needed by the error are initialized.