ast: Make AbstractFeature.qualifiedName() robust for early calls, fix #461
Disallow ambiguous types names even if they are not used as types: Constructors
with the same name and number of formal arguments will cause an error now.

Fixed duplicate type names in the base lib (`array.arrayCons` and `effect`).

For all features that declare a type, the corresponding type feature is not
created by default, independent of whether it is used or not. This avoids a lot
of problems and workarounds when creating them on demand only. Disadvantage is
that library files are slightly larger now.

`f.this.type` now work in a library module since the type features for all
children of `f` will be created.

local_mutate now prints the actual `local_mutate.this.type` in `panic`-messages.

This fixes #461 without actually changing the format of the .fum file, but by
adding all the type features. `base.fum` grows by about 10%, which is not great,
but acceptable for now.  Many file size optimizations can stil be done.