ast: Fix assignment checks for `this.type` in type features, fix #1007
The problem was that there are two representation of `this.type`: Within type
features, a type parameter contains the corresponding `this.type`, while usually
a Type instance of kind `RefOrVal.ThisType` is used. In the examle of #1007,
these two types get assigned resulting in the assignment compatibility check to
fail.

This patch fixes this be converting the type parameter variant back to the usual
`RefOrVal.ThisType` version.

This patch also adds documentation and an example to `replace_THIS_TYPE`, the
original documentation was confusing.

Also, the occasionally added `<>` after types are fixed by this patch, empty
generics must use Type.NONE, which was not the case at one place.

Finally, the toString-output for a type has been changed to include `(in type
feature)` for a `this.type` referring to the original feature's type.