ast: Improve type inference for type parameters and numeric literals, fix #1050
Numeric literals are now used as a second-class source of type information. Only
if type inference is not possible otherwise, the default type of a numeric
literal will be used.

This has the effect that in an expression `x = y`, where one side is a numeric
literal, the type parameter will be inferred from the other side.