fe: Freeze LibraryCall.actualTypeParameters(), fix #1098
Without freezing this list, a call to Clazz.actualGenerics() with these generics
will result in some of the generics being replaced, so a second call with
different actual generics will not use the correct actual generics.

This problem did not manifest for dev.flang.ast.Call since the front end calls
AbstractType.actualType on Call.actualTypeParameters, which freezes this list,
while the list in LibraryCall was not frozen.

This patch also adds a NYI-comment where Call.actualTypeParameters should freeze
the list with a reference why this is not possible yet (due to a hacky way to
obtain the result type from a lambda).

Also added one more use of List.setOrClone, just to be safe and removed a wrong
comment in LibraryCall (refering to resolveTyps, while resolution is fully done
for Libraraycall).