ast/tests: Fix #1158: type inferred from type parameter and type of type
It is not possible to obtain the type of a type feature. This type is always the
base library feature Type, so we cannot run into an endless sequence of
type-of-type-of-type...

Also, for routines and fields whose result type is inferred from a type
parameter, the type is set to the base library feature 'Type'.

The usage of base lib's 'Type' in these cases results in boxing of type values,
which allows them to be passed on (while type values themselves are unit type
values and their surrounding feature get specialized for a specific value).

Also added a test checking type values returned by routines or assigned to
fields.