fe: Create error if call lookup is ambiguous, fix #1089
Before this patch, an unqalified call `f` where `f` was found in several levels
of outer features, the innermost match was called.  Furthermore, if the
innerrmost match had an incompatible number of arguments, but an outer one
matched, an error was produced.

Now, for an assignment or unqualified call, all features with a matching name
from all outer features are determined first.  Then, this set is reduced to the
features matching the number of actual arguments.  If this set contains more
than one feature, an error is produced and it is suggested to make a qualified
call.

Additionally, the code to perform the feature lookup is simplified
significantly. The main methods used are SourceModule.lookup,
SourceModule.lookupType and FeatureAndOuter.filter.  Class FeaturesAndOuter and
method Call.calledFeatureCandidates, SourceModule.lookupFeatures have been
removed.

Error messages for ambiguous calls now contains a list of all outer contexts and
and a suggestion with sample code to perform a qualified call.

Several errors that appeared as consequential errors of ambiguous calls are
supressed now.  In particular, actual arguments to a call that cannot be
resolved are no longer checked for errors. Also added additional check()s and
made code robust against null pointer exceptions or reporting errors that are
the consequence of using Types.t_ERROR.