fe: Remove the special handling for type lookup wrt outer
Now, type lookup will cause an error when a matching type is found at different
outer levels, just the same as a call lookup would cause an error in this case.

For this to work, the special syntax `a : choice of x, y, z` was changed to no
longer require lookup of `x, y, z` such that this does not cause an error in
case of a conflict (which occured for `effectMode.abort`).

Also, the type lookup in SeouceModule.setOUterANdAddInnerForQualifiedRec had to
be fixed not to look for types in outer features unless we look for the first
type in a qualified type `a.b.c`.