fe/air/tests: Add support for a.this.type as feature result, fix #1183
A new feature `AbstractType.replace_this_Type_by_actual_outer` is used to
replace occurences of `*.this.type` in a result type of a call to `f` via
`o.f` by the actual outer types given by `o`.

Also added a regression test for this issue.

The regression test still contains parts that are commented out. These do not
work yet, but the problem seems to be independent of this commit, maybe relate
dto #1135 or some other issue. I will create a new issue for this.