util/fe/tests: Fix broken caching with mutable Lists of types as keys, fix #1192
The problem for the failing test cases was caching done by
AbstractType.actualType while the cached Type was later modified.

To solve this, first the values stored in Type._generics is now always frozen
and can no longer be mutated.

For this to be possible, a mutation of the list of generics in
Function.propagateExpectedType2 had to be removed. Now, propagateExpectedType2
no longer modifies the generic argument of the Function's result, but instead it
returns that result type and updates it in the generics list of the inherits
call to Function.

Also, Match expressions no longer mutate the choice generics of the subject's
type.

Finally, FormalGenerics.AsActuals are no longer mutated when a type using these
as generics is interned. Instead, a new instance of FormalGenerics.AsActuals is
created.  The comparison `f.generics.asActuals() == a` has been replaced by a
method `a.actualsOf(f)`.