ast: wrap arguments to call in Lazy if needed
If a called feature expects an argument of type Lazy T, but the argument
given is of type T, then the argument is first wrapped in Lazy. This is
done to avoid keeping two features around that do essentially the same
thing, such as `list` (which takes a head and a lazy tail) and
`list_of_none_lazy` (where none of the arguments are lazy).