ast: Complete rewrite of handling of outer types, fix #1180, fix #229
Types may now contain `this.type` and any outer level, e.g.,
`a.this.type.b.this.type.c.this.type.d.this.type`, or
`a.this.type.b.c.this.type.d`.

On a call, the `this.type` entries in any types, their outer types and generic
arguments are replaced by the corresponding actual type depending on the actual
outer in the call.  Method AbstractType.replace_this_type_by_actual_outer does this replacement.

New methods AbstractFeaturu.thisType()/.thisType(boolean) create the plain
`this.type` of a given feature, where all outer features are `this.type` except
for outer features for `fixed` features (that are fixed to one particular outer
feature).

New method Call.adjustThisTypeForTarget now performs all the type replacements
that depend on the target type that are required in a call.  This includes also
etype parameters for this.type in type features.

Some tests had to be updated, valWithDynOuter can now be used with the C
backend.