ast: Remove redundant checks and throws, fix #283
The first check (select >= 0 in a call to a type parameter) is handled in
resolveSelect.

The second check for that the result of gt.resolve() is not null is also
redundant since null is no longer used to indicate an error, but Types.t_ERROR
is returned in this case.