parser: Improve grammar and parser for `:` used in inherits and anonymous
The goal of this patch is to be able to parse code like

  ints(i i32) =>
    i : ints i+1

This currently fails since `i :` would be parsed as a feature declaration `i :
parent is ..`, which is fixed by no longer basing this decisions on the `:`, but
on the `is` or `=>`.

Then, `i :` would be parseed as an `anonymous`, which has very ugly syntax
anyway.  This patch changes the grammar for `anonymous` to accept only `ref :`
at the beginning, which is the way it is currently used and the only way that
actually makes sense since the result will can only be assigned to a `ref` type
of any of its ancestors.

After that, the parser will treat `i : ints i+1` as an infix operator call,
which is what we want!