tools: allow builds including reduced build information
By setting `FUZION_REPRODUCIBLE_BUILD=true` when running `make`, a build
of Fuzion can be produced which does not include the following
information: the date and time that the build was produced, the username
of the user building Fuzion as well as the hostname of the machine
building Fuzion. This is intended for perhaps reproducible builds, or
some CI environments.