ast: Fix result type inference from lazy expression, make `infix :` lazy
Moved the code to expand a call to a lazy feature `f` to `f.call` to
Call.resolveTypes such that the new, non-lazy result type will be used in type
inference.

This enables changing the last argument of `infix :` and `list(T, h t)` to be
lazy.

This permits lazy lists like the following to produce all integers starting at
`i`:

  ints(i i32) =>
    i : ints i+1