ast: Move detail error message generation from Call to AstErrors
The detail message for error "Incompatible types found during type inference for
type parameters" was partly created in Call.java and did not use the
highlighting for the types printed.

Now, the raw data, a list of positions and types, is passed to AstErrors and
converted to a String there.  Types are wrapped using `s()`.