lib/Monoid: remove `Monoid.infix ==` equality operation
Ideally, the type parameter of `Monoid` should be restricted to
`equatable` types, and the monoid property could in the future be
checked using that operation. There are few cases where using a
different equality operation would be a good idea, and in this case,
implementing your own monoid isn't that hard either.

However, we do not yet introduce a type contraint to `equatable`. This
is because in some places (`list` and `Sequence`, in particular, which
do not have an equality operation), the `Monoid` is (ab)used to
implement a nice variant of `fold` whose starting element is the neutral
element of the monoid.