ast+air: Fix actual type generation for type feature type's type parameters
Type features have an additional first type parameter, `THIS_TYPE`, and actual
type parameters for the other type parameters have to be created from this.
E.g., for

  list A

the type feature is

  list.type THIS_TYPE A

and the actual type of, e.g.,

  list.type (list (complex f32)) A

is

  list.type (list (complex f32)) (complex f32)

Where `A` was set to the type parameter of the actual type of `THIS_TYPE`.

Furthermore, the creation of type featurers requires types of the outer
features, even if these are routines that do not define a type.  Since routines
can be overloaded with different argument counts, we add the argument count to
create a unique type name (that is used, e.g., in the C backend to create type
names).