lib: introduce global min and max features
These allow conviently writing `min a b` or `max a b` and use the total
order defined on the given type.

Remove the `numeric.min` and `numeric.max` features, and replace their
usage with use of the new global features.

Rename (local) features in `ps_map` named `min` and `max` as `min0` and
`max0`. This is to avoid a conflict with the new `min` and `max`
features.