dfa: Fix quadratic performance of DFA for sequence of calls, fix #990
The problem was that a sequence of calls of the form

  a 1; a 2; a 3; a 4; a 5; ...

results in the DFA to create one new Call per iteration since the newly created
call in the current iteration was not analysed yet and hence is assumed not to
return.

This patch performs the analysis of new calls immediately unless they are
recursive or the depth of recurisve analysis of new calls exceeds 20.

The performance of the example from #990 using 128 lines of array code went down
from 22s to 2s:

before:

  > time ./build/bin/fz -c ex_990.fz -CC=echo
  ...
  real	0m22,604s
  user	0m38,519s
  sys	0m0,725s

after:

  > time ./build/bin/fz -c ex_990.fz -CC=echo
  ...
  real	0m2,241s
  user	0m8,768s
  sys	0m0,404s