parser: Fix parsing of `if` following `else`, fix #327
The problem was that an `if` statement was parsed as a single statement, not as
a block. So if an `else` was followed by a block that happened to start with an
`if`, only the `if`-statement was parsed and the remaining statements of that
block caused an error due to wrong indentation.

Now, the `else`-part is always parsed as a block.

The following code (from #327 with variations added) is legal now:

  some_fun bool is
    if true
      true
    else
      if true
        say "hello"
      else
        say "hello"
      true

  some_fun2 bool is
    if true
      true
    else if true
           say "hello"
         else
            say "hello"
         say "ok..."
         true

  some_fun3 bool is
    if true
      true
    else if true
      say "hello"
    else
      say "hello"
    say "ok..."
    true