fe: Fix for flang.dev/design/example/brd_choic.fz, redefine of choice
This fixes two things: First, it re-establishes the specifiic error produced
when an attempt is made to call a choice feature. This specific error got lost
during the recent work on call feature resolution.

Second, it adds a check that forbids the redefinition of choice features and
creates a new error if an attempt to do so is made.