parser: be lazier when creating type parameters for `choice of`, fix #1289
When creating type parameters for choice in `a : choice of x, y`, do not set the
feature references directly even though they are known, but leave the type
resolution for later. This has the effect that the outer type will be set
correctly to `outer.this.type`.

This might result in additional errors in case the feature names are
ambiguous. But if this is the case, I guess it is better to flag an error
anyway.