parser/ast: Add class OuterType to represent outer type in `choice of a, b`
Instead of using the outer feature's name when creating `choice f.a f.b`, which
could result in errors in case `f` is ambiguous, we now use a new type
`OuterType` as in `choice OuterType.a OuterType.b` that explicitly resolve to
the outer feature's type.

This simplifies the code and it might be usefuly for extending fuzion to support
`outer` as a keyword to refer to the surrounding features `this` or `this.type`.