parser: remove special cases related to escape characters
Rewrite the code to avoid special cases relating to escape characters.
In particular, the escapeChars array has been changed from an array of
`char` to an array of `int`. Two special cases remain: in `legalEscapes`
we need special handling to avoid printing a line feed and carriage
return directly. We also need special handling for carriage returns in
the input as well. Since Fuzion requires a line feed after a carriage
return, we can limit this to a minimum.