util: return END_OF_FILE in SourceFile.codePoint if pos is past EOF
This fixes an issue where a string literal at the end of file caused an
error. The problem is that the string literal is read twice, the second
time it accesses the codepoints explicitly, including the character
after the closing quote. This did not return an END_OF_FILE codepoint
until now, instead causing a precondition failure.