parser: remove the `lazy` keyword
The `lazy` keyword did not do anything, therefore it can be removed. An
issue that remains is that the boolean ternary operator can't be changed
to accept `Lazy` values due to a type inference issue. However, it seems
there is no special handling to make the ternary operator lazy, so this
should not behave differently.