fe: fix outer of Lazy being set to Lazy.call instead of surrounding feature. fixes #1358
example where this happened:
```
ex is
    a list i32 := 1:(2:nil)
```
First `nil` is wrapped in Lazy, the outer is set to ex.
Second (2:nil) is wrapped in Lazy and in the process set the outer of the first lazy even though this anonymous function was already resolved. This time to the `call` feature of the second Lazy.